move schrodinger featurizer
